+++ gaclen/examples/phong_cube/main.rs
@@ -12,6 +12,8 @@ mod geometry;
 
 use gaclen::graphics;
 
+use cgmath::One;
+
 use winit::event_loop::{ControlFlow, EventLoop};
 use winit::event::{Event, WindowEvent};
 use winit::window::WindowBuilder;
@@ -34,7 +36,7 @@ fn main() {
 	let context = graphics::context::Context::new().unwrap();
 	let device = graphics::device::Device::new(&context).unwrap();
 	println!("Initialized device: {:?}", device);
-	let mut swapchain = graphics::swapchain::Swapchain::new(&context, &device, window.clone(), graphics::swapchain::PresentMode::Immediate, graphics::PixelFormat::D16Unorm).expect("Failed to create swapchain!");
+	let mut swapchain = graphics::swapchain::Swapchain::new(&context, &device, window.clone(), graphics::swapchain::PresentMode::Immediate, graphics::image::Format::D16Unorm).expect("Failed to create swapchain!");
 
 	let albedo_pass = {
 		let vs = shaders::vertex::Shader::load(&device).unwrap();
@@ -63,7 +65,7 @@ fn main() {
         let dimensions = graphics::image::Dimensions::Dim2d { width, height };
 		let image_data = image.into_raw(); // to_rgba() returns Vec<u8> backed container
 		
-		graphics::image::create_immutable_image_from_iter(&device, image_data.iter().cloned(), dimensions, graphics::PixelFormat::R8G8B8A8Srgb).unwrap()
+		graphics::image::create_immutable_image_from_iter(&device, image_data.iter().cloned(), dimensions, graphics::image::Format::R8G8B8A8Srgb).unwrap()
 	};
 
 	let sampler = graphics::image::Sampler::simple_repeat_linear(device.logical_device());
@@ -81,19 +83,21 @@ fn main() {
 	let light_descriptor_set = Arc::new(albedo_pass.start_persistent_descriptor_set(1)
 		.add_buffer(light).unwrap()
 		.add_sampled_image(texture, sampler).unwrap()
-		.build().unwrap());
+		.build().unwrap()
+	);
 
 	let mut recreate_swapchain = false;
 
 	let mut rotation_enabled = false;
 	let mut last_x = 0;
 	let mut last_y = 0;
-	let mut object_rotation = cgmath::Quaternion::new(1.0, 0.0, 0.0, 0.0);
+	let mut object_rotation = cgmath::Quaternion::one();
 
 	// Wrap the device in a stack-allocated container to allow for temporary ownership.
 	let mut device = Some(device);
 
 	event_loop.run(move |event, _, control_flow| {
+		*control_flow = ControlFlow::Poll;
 		match event {
 			Event::WindowEvent { event: WindowEvent::CloseRequested, .. } => {
 				*control_flow = ControlFlow::Exit;
@@ -142,7 +146,8 @@ fn main() {
 		
 				let transform_descriptor_set = Arc::new(albedo_pass.start_persistent_descriptor_set(0)
 					.add_buffer(transform).unwrap()
-					.build().unwrap());
+					.build().unwrap()
+				);
 		
 				// Device ownership is taken here.
 				let frame = graphics::frame::Frame::begin(device.take().unwrap(), &swapchain).unwrap();
@@ -153,7 +158,7 @@ fn main() {
 					.build().unwrap()
 				);
 		
-				let after_frame = frame.begin_pass(&albedo_pass, framebuffer, vec![clear_color.into(), 1f32.into()])
+				let after_frame = frame.begin_pass(&albedo_pass, framebuffer, swapchain.default_viewport(), vec![clear_color.into(), 1f32.into()])
 					.draw(vec![geometry.clone()], (transform_descriptor_set, light_descriptor_set.clone()), ())
 					.finish_pass()
 				.finish();
@@ -175,8 +180,8 @@ fn main() {
 }
 
 // Ideally the view and projection matrices would be found by some 'Camera' concept.
-fn transform(rotation: cgmath::Quaternion<f32>, window_resolution: (u32, u32)) -> shaders::vertex::ty::TransformData {
-	let aspect = window_resolution.0 as f32 / window_resolution.1 as f32;
+fn transform(rotation: cgmath::Quaternion<f32>, viewport_dimensions: (u32, u32)) -> shaders::vertex::ty::TransformData {
+	let aspect = viewport_dimensions.0 as f32 / viewport_dimensions.1 as f32;
 
 	let model: cgmath::Matrix4<f32> = rotation.into();
 	let proj: cgmath::Matrix4<f32> = cgmath::PerspectiveFov { fovy: cgmath::Deg(40.0).into(), aspect, near: 0.1, far: 4.0 }.into();